1.16.4 + Fix Esmorall version mess + Other things.
On the last repo update, Esmorall deleted 1.16.1, 1.16.2 and 1.16.3 compatibility. I think this should not be like that.
I added support for this versions again and added some things to the /etf command.
