+++ EntityTrackerFixer/plugin.yml
@@ -1,6 +1,6 @@
 name: EntityTrackerFixer
 main: net.minemora.entitytrackerfixer.EntityTrackerFixer
-version: 1.3.3
+version: 1.3.4
 softdepend: [Vault]
 api-version: 1.14
 author: Esmorall
